+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- `WebCrypto` layer that satisfies the `Crypto.Crypto` requirement of `generate` using the standard Web Crypto API, so `generate` can run with `Effect.runPromise` without pulling in a platform package. Documented how to provide `Crypto` in the README. ([#2](https://github.com/just-be-dev/effect-typed-id/issues/2))
+
 ## 0.3.0
 
 ### Changed